Improve error output for synchronous instructions
Unity captures only the first logged error, so logging an error from an instruction is not good,
as it will not become the primary error.

You shouldn't be using Do in the main test without anything asynchronous before it,
so it should be fine to not log an error there.

There might be some corner cases, but you can only polish a turd (Unity Test Runner) so much...