Test Operations Status Window in Editor (#28)
Shows test operation status in an Editor window.

Supports multiple operation states, in case Responsible is used in a "re-entrant" way, e.g. a test instruction waiting for a coroutine, which yields another test instruction.

As this is using UIElements, bump minimum Unity version to 2019.4. So far I don't know of anyone using the library with 2018, but if someone is, please report an issue, and I'll look into making this conditional!